+++ pkg/plugin/settings.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/proxy"
+	"github.com/grafana/grafana-plugin-sdk-go/experimental/errorsource"
 )
 
 // Settings - data loaded from grafana settings database
@@ -52,10 +53,10 @@ const secureHeaderKeyPrefix = "secureHttpHeaders."
 
 func (settings *Settings) isValid() (err error) {
 	if settings.Host == "" {
-		return ErrorMessageInvalidHost
+		return errorsource.DownstreamError(ErrorMessageInvalidHost, false)
 	}
 	if settings.Port == 0 {
-		return ErrorMessageInvalidPort
+		return errorsource.DownstreamError(ErrorMessageInvalidPort, false)
 	}
 	return nil
 }
@@ -79,7 +80,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if port, ok := jsonData["port"].(string); ok {
 			settings.Port, err = strconv.ParseInt(port, 0, 64)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse port value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse port value: %w", err), false)
 			}
 		} else {
 			settings.Port = int64(jsonData["port"].(float64))
@@ -92,7 +93,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if secure, ok := jsonData["secure"].(string); ok {
 			settings.Secure, err = strconv.ParseBool(secure)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse secure value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse secure value: %w", err), false)
 			}
 		} else {
 			settings.Secure = jsonData["secure"].(bool)
@@ -106,7 +107,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if tlsSkipVerify, ok := jsonData["tlsSkipVerify"].(string); ok {
 			settings.InsecureSkipVerify, err = strconv.ParseBool(tlsSkipVerify)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse tlsSkipVerify value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse tlsSkipVerify value: %w", err), false)
 			}
 		} else {
 			settings.InsecureSkipVerify = jsonData["tlsSkipVerify"].(bool)
@@ -116,7 +117,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if tlsAuth, ok := jsonData["tlsAuth"].(string); ok {
 			settings.TlsClientAuth, err = strconv.ParseBool(tlsAuth)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse tlsAuth value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse tlsAuth value: %w", err), false)
 			}
 		} else {
 			settings.TlsClientAuth = jsonData["tlsAuth"].(bool)
@@ -126,7 +127,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if tlsAuthWithCACert, ok := jsonData["tlsAuthWithCACert"].(string); ok {
 			settings.TlsAuthWithCACert, err = strconv.ParseBool(tlsAuthWithCACert)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse tlsAuthWithCACert value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse tlsAuthWithCACert value: %w", err), false)
 			}
 		} else {
 			settings.TlsAuthWithCACert = jsonData["tlsAuthWithCACert"].(bool)
@@ -174,7 +175,7 @@ func LoadSettings(ctx context.Context, config backend.DataSourceInstanceSettings
 		if forwardGrafanaHeaders, ok := jsonData["forwardGrafanaHeaders"].(string); ok {
 			settings.ForwardGrafanaHeaders, err = strconv.ParseBool(forwardGrafanaHeaders)
 			if err != nil {
-				return settings, fmt.Errorf("could not parse forwardGrafanaHeaders value: %w", err)
+				return settings, errorsource.DownstreamError(fmt.Errorf("could not parse forwardGrafanaHeaders value: %w", err), false)
 			}
 		} else {
 			settings.ForwardGrafanaHeaders = jsonData["forwardGrafanaHeaders"].(bool)